+++ tests/architecture/NativeUniversalBalance/functions/MultiDepositNativeFor.t.sol
@@ -238,7 +238,7 @@ contract MultiDepositNativeForTest is TestBaseNativeUniversalBalance {
     ) {
         for (uint256 i; i < 3; i++) {
             vm.assume(
-                0 < amounts_[i] && amounts_[i] < type(uint216).max / _ONE
+                0 < amounts_[i] && amounts_[i] < type(uint256).max / _ONE
             );
 
             amounts.push(amounts_[i]);
