+++ tests/architecture/NativeUniversalBalance/functions/Withdraw.t.sol
@@ -47,7 +47,7 @@ contract NativeUniversalBalanceWithdrawTest is TestBaseNativeUniversalBalance {
     function test_nativeUniversalBalanceWithdraw_fail_whenExceedsLentBalance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
 
         deal(user1, amount);
 
@@ -64,7 +64,7 @@ contract NativeUniversalBalanceWithdrawTest is TestBaseNativeUniversalBalance {
     function test_nativeUniversalBalanceWithdraw_fail_whenExceedsSittingBalance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
 
         deal(user1, amount);
 
@@ -93,7 +93,7 @@ contract NativeUniversalBalanceWithdrawTest is TestBaseNativeUniversalBalance {
         uint256 withdrawAmount
     ) public {
         vm.assume(
-            0 < depositAmount && depositAmount < type(uint216).max / _ONE
+            0 < depositAmount && depositAmount < type(uint256).max / _ONE
         );
         vm.assume(0 < withdrawAmount && withdrawAmount <= depositAmount);
 
@@ -139,7 +139,7 @@ contract NativeUniversalBalanceWithdrawTest is TestBaseNativeUniversalBalance {
         uint256 withdrawAmount
     ) public {
         vm.assume(
-            0 < depositAmount && depositAmount < type(uint216).max / _ONE
+            0 < depositAmount && depositAmount < type(uint256).max / _ONE
         );
         vm.assume(0 < withdrawAmount && withdrawAmount <= depositAmount);
